requirements.txt
opencv-python
mediapipe
numpy
Pillow
matplotlib
